Add Clanker's Cavern (aquatic)
History was not tracked for this song.

There are two versions: one that has a higher pitch than the original for the actual melody, and one that matches the original.  I think the inaccurate one sounds better.

This song was started on October ‎1, ‎2016 at 3:37 PM and finished on December ‎30, ‎2016 at ‎5:34 PM.