Add rusty bucket bay engine room
As with before, history was not tracked.

This song was started on ‎October ‎13, ‎2016 at ‏‎6:06 PM and finished on ‎October ‎16, ‎2016 at 8:43 PM.  The ins2 version was created on ‎December ‎17, ‎2016 at 8:33 PM.

Sound effects were created by injecting a custom midi into the ROM and then recording the playback.  The midi in question is included, along with the source for it (sounds wonderful, doesn't it?)