Add Clanker's Cavern - Inside Clanker - Aquatic
This is one of my favorite songs in the game (both the aquatic version and regular version), just for how it constantly changes but also stays the same.  The only thing I don't feel happy with for this version of it is the chords at the end -- but I can't use famitracker's apreigo with aquatic songs (it sounds bad) and adding more channels ONLY for that bit seems silly.